refactor(serializer): #4400 pack_common_attr_bits를 document_core::converters에서 serializer로 이동
CTRL_HEADER attr 비트를 CommonObjAttr 필드로부터 합성하는 pack_common_attr_bits는
본질적으로 직렬화 로직인데 document_core/converters/common_obj_attr_writer.rs에
있었고, 정작 진짜 직렬화기(serializer/control.rs, serializer/hml/preflight.rs)가
이를 역으로 import하는 역방향 의존이 있었다. document_core/converters/mod.rs 자신의
헌장("잘 작동하는 직렬화기 어깨 위에 서자 — 직렬화기 자체를 수정하지 않고 IR만
정렬한다")과도 어긋난다.

pack_common_attr_bits와 그 8개 헬퍼(_to_bits 함수들)를 serializer/control.rs로
옮기고 pub(crate)로 노출했다. document_core 쪽 3개 호출부(converters/hwpx_to_hwp.rs,
converters/common_obj_attr_writer.rs의 serialize_common_obj_attr/sync_anchor_bits,
commands/object_ops/common.rs)는 새 위치를 import하도록만 바꿨다 — document_core가
serializer를 참조하는 것은 이미 여러 곳(field_query.rs, document.rs 등)에 있는
정상적인 방향이다.

순수 이동이며 로직 변경 없음. 관련 유닛 테스트(common_obj_attr_writer 15개,
serializer::control 25개) 무변경 통과 + 전체 스위트(477 테스트 바이너리) 통과로
동작 불변을 확인했다(issue_2007_nested_cell_pagination의 서로 다른 서브테스트가
전체 스위트 실행마다 다르게 실패하지만 격리 실행에서는 항상 통과 — 부하 하 flaky,
본 변경과 무관).
